KAFKA-17175: Remove BrokerNode and ControllerNode
