--- core/src/test/java/kafka/testkit/BrokerNodeTest.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package kafka.testkit;
-
-import org.apache.kafka.common.Uuid;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class BrokerNodeTest {
-
-    @Test
-    public void testInvalidBuilder() {
-        Assertions.assertEquals("You must set the node id.",
-                Assertions.assertThrows(IllegalArgumentException.class, () -> BrokerNode.builder()
-                        .setBaseDirectory("foo")
-                        .setClusterId(Uuid.randomUuid().toString())
-                        .build()).getMessage());
-
-        Assertions.assertEquals("The value of numLogDirectories should be at least 1.",
-                Assertions.assertThrows(IllegalArgumentException.class, () -> BrokerNode.builder()
-                        .setBaseDirectory("foo")
-                        .setClusterId(Uuid.randomUuid().toString())
-                        .setId(0)
-                        .setNumLogDirectories(0)
-                        .build()).getMessage());
-    }
-}
